add park to Itinerary Preview
